accessible/tests/mochitest/focus/test_takeFocus.html

<html>

<head>
  <title>nsIAccessible::takeFocus testing</title>

  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>
  <script type="application/javascript"
          src="../states.js"></script>
  <script type="application/javascript"
          src="../events.js"></script>

  <script type="application/javascript">

    async function takeFocusInvoker(aID) {
      let accessible = getAccessible(aID);
      let p = waitForEvent(EVENT_FOCUS, accessible);
      accessible.takeFocus();
      let evt = await p;
      testStates(evt.accessible, STATE_FOCUSED);
    }

    async function doTest() {
      disableLogging(); // from test_focusedChild

      await takeFocusInvoker("aria-link");
      await takeFocusInvoker("aria-link2");
      await takeFocusInvoker("link");
      await takeFocusInvoker(document);
      await takeFocusInvoker("lb_item2");
      await takeFocusInvoker(document);
      await takeFocusInvoker("lb_item3.2");
      await takeFocusInvoker(document);
      await takeFocusInvoker("lb_item3.1");

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>

<body>

  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=429547"
     title="Support aria-activedescendant usage in nsIAccesible::TakeFocus()">
    Mozilla Bug 429547
  </a>
  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=452710"
     title="nsIAccessible::takeFocus testing">
    Mozilla Bug 452710
  </a>
  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=706067"
     title="Make takeFocus work on widget items">
    Mozilla Bug 706067
  </a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <span id="aria-link" role="link" tabindex="0">link</span>
  <span id="aria-link2" role="link" tabindex="0">link</span>

  <a id="link" href="">link</a>

  <div role="listbox" aria-activedescendant="item1" id="container" tabindex="1">
    <div role="option" id="item1">item1</div>
    <div role="option" id="item2">item2</div>
    <div role="option" id="item3">item3</div>
  </div>

  <select id="listbox" size="5">
    <option id="lb_item1">item1</option>
    <option id="lb_item2">item2</option>
    <optgroup>
      <option id="lb_item3.1">item 3.1</option>
      <option id="lb_item3.2">item 3.2</option>
    </optgroup>
  </select>
</body>
</html>
